+++ sync-for-commerce/src/codatsynccommerce/models/shared/__init__.py
@@ -17,7 +17,6 @@ from .customer import *
 from .dataconnectionerror import *
 from .dataconnectionstatus import *
 from .datatypefeature import *
-from .daterange import *
 from .errormessage import *
 from .featurestate import *
 from .featuretype import *
@@ -45,6 +44,7 @@ from .sourcetype import *
 from .supportedfeature import *
 from .syncflowurl import *
 from .syncrange import *
+from .syncstatus import *
 from .syncsummary import *
 from .synctolatestargs import *
 from .taxrateamount import *
@@ -52,4 +52,4 @@ from .taxratemapping import *
 from .updateconnection import *
 from .visibleaccounts import *
 
-__all__ = ["AccountOption","Branding","BrandingButton","BrandingImage","BrandingLogo","Companies","Company","CompanySyncStatus","ConfigAccount","Configuration","Connection","ConnectionSourceType","Connections","CreateCompany","Customer","DataConnectionError","DataConnectionStatus","DataTypeFeature","DataTypeFeatureDataType","DateRange","ErrorMessage","FeatureState","FeatureType","Fees","FeesSupplier","Grouping","GroupingLevels","GroupingPeriod","HalRef","ImageReference","Integration","Integrations","InvoiceLevelSelection","InvoiceLineLevelSelection","InvoiceStatus","Links","Localization","NewPayments","NewTaxRates","Option","Payments","Sales","Security","SourceType","SupportedFeature","SyncFlowURL","SyncRange","SyncRangeDateRange","SyncSummary","SyncToLatestArgs","TaxRateAmount","TaxRateMapping","UpdateConnection","VisibleAccounts"]
+__all__ = ["AccountOption","Branding","BrandingButton","BrandingImage","BrandingLogo","Companies","Company","CompanySyncStatus","ConfigAccount","Configuration","Connection","ConnectionSourceType","Connections","CreateCompany","Customer","DataConnectionError","DataConnectionStatus","DataTypeFeature","DataTypeFeatureDataTypes","ErrorMessage","FeatureState","FeatureType","Fees","FeesSupplier","Grouping","GroupingLevels","GroupingPeriod","HalRef","ImageReference","Integration","Integrations","InvoiceLevelSelection","InvoiceLineLevelSelection","InvoiceStatus","Links","Localization","NewPayments","NewTaxRates","Option","Payments","Sales","Security","SourceType","SupportedFeature","SyncFlowURL","SyncRange","SyncRangeDateRange","SyncStatus","SyncSummary","SyncSummarySyncDateRangeUtc","SyncToLatestArgs","TaxRateAmount","TaxRateMapping","UpdateConnection","VisibleAccounts"]