+++ sync-for-commerce/src/codatsynccommerce/models/shared/visibleaccounts.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class VisibleAccounts:
-    r"""Success"""
     visible_accounts: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('visibleAccounts'), 'exclude': lambda f: f is None }})
     r"""Visible accounts on sync flow."""